New translations update tasks (#91)
* Fix and refactor solidus_i18n:update_default task

- The file download was not working
- Download the file under /config/locales

* Add i18n-tasks development dependency

This gem can replace our custom, old and untested tasks for Yaml files
management.
This is not compatible with older solidus versions but that's not a problem. We
only need this when using the master branch.

* Use i18n-tasks for YAML files update

Remove the old code of the solidus_i18n:sync task and document the use of
i18n-tasks.