Do not configure the fallback locale
This is a setting the host app should be making.

This is what it does: It adds the `default_locale` as
a fallback locale to all `available_locales`. This means
that for a store with e.g. a German default locale and en English
locale, missing English translations would be replaced by
their German equivalents rather than show up as missing
translations.

This fallback locale feature is very nice, but not very well
documented, and should IMO not be triggered from any gem's
initializer.